+++ source/health-check/tables/insert-into-dbo.appInternalTasks.sql
@@ -19,7 +19,8 @@ INSERT	INTO [dbo].[appInternalTasks] ([id], [descriptor], [task_name], [flg_acti
 				SELECT   65536, 'dbo.usp_hcCollectErrorlogMessages', 'Collect SQL Server errorlog Messages', NULL, 3 AS [priority] UNION ALL
 				SELECT  131072, 'dbo.usp_hcCollectOSEventLogs', 'Collect OS Event Logs', NULL, 4 AS [priority] UNION ALL
 				SELECT  262144, 'dbo.usp_hcCollectSQLServerAgentJobsStatus', 'Collect SQL Server Agent Jobs Status', NULL, 5 AS [priority] UNION ALL
-				SELECT  524288, 'dbo.usp_hcCollectEventMessages', 'Collect Internal Event Messages', NULL , 6 AS [priority]
+				SELECT  524288, 'dbo.usp_hcCollectEventMessages', 'Collect Internal Event Messages', NULL , 6 AS [priority] UNION ALL
+				SELECT 8388608, 'dbo.usp_hcCollectDatabaseGrowth', 'Collect Database Growth Events', NULL , 7 AS [priority]
 			)S
 		LEFT JOIN [dbo].[appInternalTasks] ait ON S.[id] = ait.[id]
 		WHERE ait.[id] IS NULL